Refactor into separate go files: comment, post, & web handlers
