+++ web/handler.go
@@ -3,7 +3,6 @@ package server
 import (
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
-	"github.com/google/uuid"
 	"github.com/rojerdu-dev/gothreadit"
 	"html/template"
 	"net/http"
@@ -15,22 +14,26 @@ func NewHandler(store gothreadit.Store) *Handler {
 		store: store,
 	}
 
+	threads := ThreadHandler{store}
+	posts := PostHandler{store}
+	comments := CommentsHandler{store}
+
 	h.Use(middleware.Logger)
 
 	h.Get("/", h.Home())
 	h.Route("/threads", func(r chi.Router) {
-		r.Get("/", h.ThreadsList())
-		r.Get("/new", h.ThreadsCreate())
-		r.Post("/", h.ThreadsStore())
-		r.Get("/{id}", h.ThreadsShow())
-		r.Post("/{id}", h.ThreadsDelete())
-		r.Get("/{id}/new", h.PostsCreate())
-		r.Post("/{id}", h.PostsStore())
-		r.Get("/{threadID}/{postID}", h.PostsShow())
-		r.Get("/{threadID}/{postID}/vote", h.PostsVote())
-		r.Post("/{threadID}/{postID}", h.CommentsStore())
+		r.Get("/", threads.List())
+		r.Get("/new", threads.Create())
+		r.Post("/", threads.Store())
+		r.Get("/{id}", threads.Show())
+		r.Post("/{id}", threads.Delete())
+		r.Get("/{id}/new", posts.Create())
+		r.Post("/{id}", posts.Store())
+		r.Get("/{threadID}/{postID}", posts.Show())
+		r.Get("/{threadID}/{postID}/vote", posts.Vote())
+		r.Post("/{threadID}/{postID}", comments.Store())
 	})
-	h.Get("/comments/{id}/vote", h.CommentsVote())
+	h.Get("/comments/{id}/vote", comments.Vote())
 
 	h.Get("/html", func(w http.ResponseWriter, r *http.Request) {
 		t := template.Must(template.New("layout.html").ParseGlob("templates/includes/*.html"))
@@ -78,294 +81,5 @@ func (h *Handler) Home() http.HandlerFunc {
 		}
 
 		tmpl.Execute(w, data{pp})
-	}
-}
-
-func (h *Handler) ThreadsList() http.HandlerFunc {
-	type data struct {
-		Threads []gothreadit.Thread
-	}
-	tmpl := template.Must(template.ParseFiles("templates/layout.html", "templates/threads.html"))
-	return func(w http.ResponseWriter, r *http.Request) {
-		tt, err := h.store.Threads()
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		tmpl.Execute(w, data{Threads: tt})
-	}
-}
-
-func (h *Handler) ThreadsCreate() http.HandlerFunc {
-	tmpl := template.Must(template.ParseFiles("templates/layout.html", "templates/thread_create.html"))
-	return func(w http.ResponseWriter, r *http.Request) {
-		tmpl.Execute(w, nil)
-	}
-}
-
-func (h *Handler) ThreadsShow() http.HandlerFunc {
-	type data struct {
-		Thread gothreadit.Thread
-		Posts  []gothreadit.Post
-	}
-
-	tmpl := template.Must(template.ParseFiles("templates/layout.html", "templates/thead.html"))
-	return func(w http.ResponseWriter, r *http.Request) {
-		idStr := chi.URLParam(r, "id")
-		id, err := uuid.Parse(idStr)
-		if err != nil {
-			http.NotFound(w, r)
-			return
-		}
-		t, err := h.store.Thread(id)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-		pp, err := h.store.PostsByThread(t.ID)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-		}
-		tmpl.Execute(w, data{t, pp})
-	}
-}
-
-func (h *Handler) ThreadsStore() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		title := r.FormValue("title")
-		description := r.FormValue("description")
-
-		err := h.store.CreateThread(&gothreadit.Thread{
-			uuid.New(),
-			title,
-			description,
-		})
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		http.Redirect(w, r, "/threads", http.StatusFound)
-	}
-}
-
-func (h *Handler) ThreadsDelete() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		idStr := chi.URLParam(r, "id")
-		id, err := uuid.Parse(idStr)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-		}
-
-		err = h.store.DeleteThread(id)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		http.Redirect(w, r, "/threads", http.StatusFound)
-	}
-}
-
-func (h *Handler) PostsCreate() http.HandlerFunc {
-	type data struct {
-		Thread gothreadit.Thread
-	}
-
-	tmpl := template.Must(template.ParseFiles("templates/layout.html", "templates/post_create.html"))
-	return func(w http.ResponseWriter, r *http.Request) {
-		idStr := chi.URLParam(r, "id")
-
-		id, err := uuid.Parse(idStr)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		t, err := h.store.Thread(id)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		tmpl.Execute(w, data{t})
-	}
-}
-
-func (h *Handler) PostsShow() http.HandlerFunc {
-	type data struct {
-		Thread   gothreadit.Thread
-		Post     gothreadit.Post
-		Comments []gothreadit.Comment
-	}
-
-	tmpl := template.Must(template.ParseFiles("templates/layout.html", "templates/post.html"))
-	return func(w http.ResponseWriter, r *http.Request) {
-		postIDStr := chi.URLParam(r, "postID")
-		threadIDStr := chi.URLParam(r, "threadID")
-
-		postID, err := uuid.Parse(postIDStr)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		threadID, err := uuid.Parse(threadIDStr)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		p, err := h.store.Post(postID)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		cc, err := h.store.CommentsByPost(p.ID)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		t, err := h.store.Thread(threadID)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		tmpl.Execute(w, data{
-			Thread:   t,
-			Post:     p,
-			Comments: cc,
-		})
-	}
-}
-
-func (h *Handler) PostsStore() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		title := r.FormValue("title")
-		content := r.FormValue("content")
-
-		idStr := chi.URLParam(r, "id")
-
-		id, err := uuid.Parse(idStr)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		t, err := h.store.Thread(id)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		p := &gothreadit.Post{
-			ID:       uuid.New(),
-			ThreadID: t.ID,
-			Title:    title,
-			Content:  content,
-		}
-
-		err = h.store.CreatePost(p)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		http.Redirect(w, r, "/threads", http.StatusFound)
-	}
-}
-
-func (h *Handler) PostsVote() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		idStr := chi.URLParam(r, "postID")
-
-		id, err := uuid.Parse(idStr)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		p, err := h.store.Post(id)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		dir := r.URL.Query().Get("dir")
-		if dir == "up" {
-			p.Votes++
-		} else if dir == "down" {
-			p.Votes--
-		}
-
-		err = h.store.UpdatePost(&p)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		http.Redirect(w, r, r.Referer(), http.StatusFound)
-	}
-}
-
-func (h *Handler) CommentsStore() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		content := r.FormValue("content")
-		idStr := chi.URLParam(r, "PostID")
-
-		id, err := uuid.Parse(idStr)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		c := &gothreadit.Comment{
-			ID:      uuid.New(),
-			PostID:  id,
-			Content: content,
-		}
-		err = h.store.CreateComment(c)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		http.Redirect(w, r, r.Referer(), http.StatusFound)
-	}
-}
-
-func (h *Handler) CommentsVote() http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		idStr := chi.URLParam(r, "id")
-
-		id, err := uuid.Parse(idStr)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		c, err := h.store.Comment(id)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		dir := r.URL.Query().Get("dir")
-		if dir == "up" {
-			c.Votes++
-		} else if dir == "down" {
-			c.Votes--
-		}
-
-		err = h.store.UpdateComment(&c)
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		http.Redirect(w, r, r.Referer(), http.StatusFound)
 	}
 }
