+++ web/web_handler.go
@@ -1,0 +1,104 @@
+package server
+
+import (
+	"github.com/go-chi/chi/v5"
+	"github.com/google/uuid"
+	"github.com/rojerdu-dev/gothreadit"
+	"net/http"
+)
+
+type ThreadHandler struct {
+	store gothreadit.Store
+}
+
+// ThreadsList
+// ThreadsCreate
+// ThreadsStore
+// ThreadsShow
+// ThreadsDelete
+
+func (th *ThreadHandler) List() http.HandlerFunc {
+	type data struct {
+		Threads []gothreadit.Thread
+	}
+	tmpl := template.Must(template.ParseFiles("templates/layout.html", "templates/threads.html"))
+	return func(w http.ResponseWriter, r *http.Request) {
+		tt, err := th.store.Threads()
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+		tmpl.Execute(w, data{Threads: tt})
+	}
+}
+
+func (th *ThreadHandler) Create() http.HandlerFunc {
+	tmpl := template.Must(template.ParseFiles("templates/layout.html", "templates/thread_create.html"))
+	return func(w http.ResponseWriter, r *http.Request) {
+		tmpl.Execute(w, nil)
+	}
+}
+
+func (th *ThreadHandler) Show() http.HandlerFunc {
+	type data struct {
+		Thread gothreadit.Thread
+		Posts  []gothreadit.Post
+	}
+
+	tmpl := template.Must(template.ParseFiles("templates/layout.html", "templates/thead.html"))
+	return func(w http.ResponseWriter, r *http.Request) {
+		idStr := chi.URLParam(r, "id")
+		id, err := uuid.Parse(idStr)
+		if err != nil {
+			http.NotFound(w, r)
+			return
+		}
+		t, err := th.store.Thread(id)
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+		pp, err := th.store.PostsByThread(t.ID)
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+		}
+		tmpl.Execute(w, data{t, pp})
+	}
+}
+
+func (th *ThreadHandler) Store() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		title := r.FormValue("title")
+		description := r.FormValue("description")
+
+		err := th.store.CreateThread(&gothreadit.Thread{
+			uuid.New(),
+			title,
+			description,
+		})
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+
+		http.Redirect(w, r, "/threads", http.StatusFound)
+	}
+}
+
+func (th *ThreadHandler) Delete() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		idStr := chi.URLParam(r, "id")
+		id, err := uuid.Parse(idStr)
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusBadRequest)
+		}
+
+		err = th.store.DeleteThread(id)
+		if err != nil {
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+
+		http.Redirect(w, r, "/threads", http.StatusFound)
+	}
+}
